+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This project adheres to [Semantic Versioning](http://semver.org)
 
 ## [2.4.2] - 2024-08-12
 ### Added
-- added "|" to be the default scan delimeter for TF version 4
+- added "|" to be the default scan delimiter for TF version 4
 
 ## [2.4.0] - 2024-04-10
 ### Added